[client][#8]customControl 구조개선
