+++ client/src/lib/naverMapV2/hexagonSelectorHandler/index.js
@@ -1,77 +1,7 @@
 import customControlGroupHandler from '../customControlGroupHandler';
-import utils from '@/lib/utils';
 
-const getStyle = (focus, selected) => ([
-  focus ? 'background: #BDBDBD;' : 'background: #EEEEEE;',
-  selected ? 'color: black;' : 'color: grey;',
-  'padding: 10px;',
-  'margin: 4px;',
-].join(''));
-
-/*
- * 지도 위에 고정된 위치에 제공되는 버튼(CustomControl)
- *
- * 관련정보
- * https://navermaps.github.io/maps.js.ncp/docs/tutorial-4-control-custom-p1.example.html#
- * https://navermaps.github.io/maps.js.ncp/docs/naver.maps.CustomControl.html
- *
-*/
-const getHtml = (elementStatusMap) => {
-  const html = utils.convertObjValuesToList(elementStatusMap).map((v) => {
-    const {
-      key,
-      focus,
-      selected,
-      meta: {
-        name,
-      },
-    } = v;
-    const style = getStyle(focus, selected);
-    return `<div class="${key}" style="${style}">${name}</div>`;
-  });
-
-  const style = [
-    'background: white;',
-    'padding: 1px;',
-    'margin: 10px;',
-    'border: 1px solid black;',
-  ].join('');
-
-  return [
-    `<div style="${style}">`,
-    ...html,
-    '</div>',
-  ].join('');
-};
-
-const getElementStatusMapInitialized = () => ({
-  point: {
-    key: 'point',
-    focus: false,
-    selected: false,
-    meta: {
-      name: '점',
-    },
-  },
-  polyline: {
-    key: 'polyline',
-    focus: false,
-    selected: false,
-    meta: {
-      name: '직선',
-    },
-  },
-  polygon: {
-    key: 'polygon',
-    focus: false,
-    selected: false,
-    meta: {
-      name: '폴리곤',
-    },
-  },
-});
 class HexagonSelector {
-  #customControlButtonGroup
+  #hexagonSelectorButtonGroup
 
   #meta
 
@@ -83,9 +13,7 @@ class HexagonSelector {
     // eslint-disable-next-line max-len
     this.#meta = meta;
     // eslint-disable-next-line max-len
-    this.#customControlButtonGroup = customControlGroupHandler.createCustomControlGroup({
-      elementStatusMap: getElementStatusMapInitialized(),
-      onChangeHtml: getHtml,
+    this.#hexagonSelectorButtonGroup = customControlGroupHandler.createCustomControlHexagonSelectorButtonGroup({
       meta: this.#meta,
     });
   }
@@ -105,8 +33,7 @@ class HexagonSelector {
       return;
     }
     this.#map = map;
-    this.#customControlButtonGroup.setNaverMap(this.#map);
-    this.#customControlButtonGroup.draw(this.#map);
+    this.#hexagonSelectorButtonGroup.setNaverMap(this.#map);
   }
 
   /**
@@ -116,8 +43,8 @@ class HexagonSelector {
    * @return {void} 반환값 없음
    */
   remove() {
-    if (this.#customControlButtonGroup) {
-      this.#customControlButtonGroup.remove();
+    if (this.#hexagonSelectorButtonGroup) {
+      this.#hexagonSelectorButtonGroup.remove();
     }
   }
 
@@ -128,10 +55,10 @@ class HexagonSelector {
    */
   destroy() {
     this.remove();
-    if (this.#customControlButtonGroup) {
-      this.#customControlButtonGroup.destroy();
+    if (this.#hexagonSelectorButtonGroup) {
+      this.#hexagonSelectorButtonGroup.destroy();
     }
-    this.#customControlButtonGroup = null;
+    this.#hexagonSelectorButtonGroup = null;
 
     this.#meta = null;
     this.#map = null;
